+++ soane/comms/__init__.py
@@ -2,6 +2,7 @@
 Package definition for 'soane.comms'.
 '''
 
-from soane.comms._base import group
-from soane.comms.list  import list_
-from soane.comms.read  import read
+from soane.comms._base  import group
+from soane.comms.create import create
+from soane.comms.list   import list_
+from soane.comms.read   import read